Fixed backwards-compatibility breaking issue with string interpolation
The new implementation of string interpolation could lead to breaking backwards-compatibility with older versions, if the string included an escape character followed by an interpolation end sentinel. This led to just the interpolation end sentinel to be printed, which has now been fixed, by printing the escape character as well.